+++ projects/angular2-smart-table/src/lib/components/filter/filter.module.ts
@@ -9,6 +9,7 @@ import {InputFilterComponent} from './filter-types/input-filter.component';
 import {SelectFilterComponent} from './filter-types/select-filter.component';
 import {DefaultFilter} from './filter-types/default-filter';
 import {FilterDefault} from './filter-default';
+import {MultiSelectFilterComponent} from './filter-types/multiselect-filter.component';
 
 const FILTER_COMPONENTS = [
   FilterDefault,
@@ -19,7 +20,9 @@ const FILTER_COMPONENTS = [
   CheckboxFilterComponent,
   InputFilterComponent,
   SelectFilterComponent,
+  MultiSelectFilterComponent,
 ];
+
 
 @NgModule({
   imports: [