+++ projects/demo/src/app/pages/examples/examples.module.ts
@@ -1,7 +1,7 @@
 import {NgModule} from '@angular/core';
 import {RouterModule} from '@angular/router';
 import {CommonModule} from '@angular/common';
-import {HttpClientModule} from '@angular/common/http';
+import {provideHttpClient} from "@angular/common/http";
 import {Angular2SmartTableModule} from 'angular2-smart-table';
 
 import {SharedModule} from '../../shared/shared.module';
@@ -65,7 +65,6 @@ const EXAMPLES_COMPONENTS = [
 @NgModule({
   imports: [
     CommonModule,
-    HttpClientModule,
     RouterModule.forChild(routes),
     Angular2SmartTableModule,
     SharedModule,
@@ -74,5 +73,6 @@ const EXAMPLES_COMPONENTS = [
     ExamplesComponent,
     ...EXAMPLES_COMPONENTS,
   ],
+  providers: [provideHttpClient()]
 })
 export class ExamplesModule { }